+++ R/000-wrappers.R
@@ -4052,6 +4052,12 @@ class(`PlRExpr`) <- c("PlRExpr__bundle", "savvy_polars__sealed")
   }
 }
 
+`PlRLazyFrame_sink_batches` <- function(self) {
+  function(`lambda`, `maintain_order`, `chunk_size` = NULL) {
+    .savvy_wrap_PlRLazyFrame(.Call(savvy_PlRLazyFrame_sink_batches__impl, `self`, `lambda`, `maintain_order`, `chunk_size`))
+  }
+}
+
 `PlRLazyFrame_sink_csv` <- function(self) {
   function(`target`, `include_bom`, `include_header`, `separator`, `line_terminator`, `quote_char`, `batch_size`, `retries`, `sync_on_close`, `maintain_order`, `mkdir`, `decimal_comma`, `datetime_format` = NULL, `date_format` = NULL, `time_format` = NULL, `float_scientific` = NULL, `float_precision` = NULL, `null_value` = NULL, `quote_style` = NULL, `storage_options` = NULL) {
     .savvy_wrap_PlRLazyFrame(.Call(savvy_PlRLazyFrame_sink_csv__impl, `self`, `target`, `include_bom`, `include_header`, `separator`, `line_terminator`, `quote_char`, `batch_size`, `retries`, `sync_on_close`, `maintain_order`, `mkdir`, `decimal_comma`, `datetime_format`, `date_format`, `time_format`, `float_scientific`, `float_precision`, `null_value`, `quote_style`, `storage_options`))
@@ -4213,6 +4219,7 @@ class(`PlRExpr`) <- c("PlRExpr__bundle", "savvy_polars__sealed")
   e$`serialize_binary` <- `PlRLazyFrame_serialize_binary`(ptr)
   e$`serialize_json` <- `PlRLazyFrame_serialize_json`(ptr)
   e$`shift` <- `PlRLazyFrame_shift`(ptr)
+  e$`sink_batches` <- `PlRLazyFrame_sink_batches`(ptr)
   e$`sink_csv` <- `PlRLazyFrame_sink_csv`(ptr)
   e$`sink_ipc` <- `PlRLazyFrame_sink_ipc`(ptr)
   e$`sink_json` <- `PlRLazyFrame_sink_json`(ptr)